+++ LuxTracer/LuxTracer/src/Ray.cpp
@@ -2,11 +2,11 @@
 
 namespace lux
 {
-	Ray::Ray(math::dvec3 const& pos, math::dvec3 const& dir)
+	Ray::Ray(math::vec3 const& pos, math::vec3 const& dir)
 		: position{ pos }, direction{ dir }
 	{}
 
-	math::dvec3 const Ray::At(double t) const noexcept
+	math::vec3 const Ray::At(float t) const noexcept
 	{
 		return position + direction * t;
 	}
